change tinyint type=>Integer to Byte

+++ src/main/java/com/thestar/member/entity/MemberVO.java
@@ -35,14 +35,14 @@ public class MemberVO implements Serializable {
     private LocalDate memberBirthday; // 欄位為 DATE，建議用 LocalDate
 
     @Column(name = "MEMBER_GENDER")
-    private Integer memberGender; // 0:FEMALE, 1:MALE, 2:UNKNOWN
+    private Byte memberGender; // 0:FEMALE, 1:MALE, 2:UNKNOWN
 
     @Lob
     @Column(name = "MEMBER_PICTURE", columnDefinition = "LONGBLOB")
     private byte[] memberPicture; // 對應 LONGBLOB 欄位
 
     @Column(name = "MEMBER_STATUS", nullable = false)
-    private Integer memberStatus = 0; // 預設 0:NOT ENABLED, 1:ENABLED, 2:DISENABLED
+    private Byte memberStatus = 0; // 預設 0:NOT ENABLED, 1:ENABLED, 2:DISENABLED
 
     @Column(name = "CREATED_TIME", nullable = false)
     private LocalDateTime createdTime;
@@ -117,11 +117,11 @@ public class MemberVO implements Serializable {
         this.memberBirthday = memberBirthday;
     }
 
-    public Integer getMemberGender() {
+    public Byte getMemberGender() {
         return memberGender;
     }
 
-    public void setMemberGender(Integer memberGender) {
+    public void setMemberGender(Byte memberGender) {
         this.memberGender = memberGender;
     }
 
@@ -133,11 +133,11 @@ public class MemberVO implements Serializable {
         this.memberPicture = memberPicture;
     }
 
-    public Integer getMemberStatus() {
+    public Byte getMemberStatus() {
         return memberStatus;
     }
 
-    public void setMemberStatus(Integer memberStatus) {
+    public void setMemberStatus(Byte memberStatus) {
         this.memberStatus = memberStatus;
     }
